backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
httpx>=0.27.0
Pillow>=10.0.0
imagehash>=4.3.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
python-multipart>=0.0.9
pillow-heif>=0.14.0
websockets>=13.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
